moved project into new package structure
