+++ Stride.CommunityToolkit.sln
@@ -79,6 +79,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Example12_Particles", "exam
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Example13_RootRendererShader", "examples\code-only\Example13_RootRendererShader\Example13_RootRendererShader\Example13_RootRendererShader.csproj", "{E8FA4D0B-D5B3-48B4-AFB5-CEF86BFF88BC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Stride.CommunityToolkit.DebugShapes", "src\Stride.CommunityToolkit.DebugShapes\Stride.CommunityToolkit.DebugShapes.csproj", "{11BB58D0-9592-424C-B9C9-71D545BFABBB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -208,6 +210,10 @@ Global
 		{E8FA4D0B-D5B3-48B4-AFB5-CEF86BFF88BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8FA4D0B-D5B3-48B4-AFB5-CEF86BFF88BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E8FA4D0B-D5B3-48B4-AFB5-CEF86BFF88BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{11BB58D0-9592-424C-B9C9-71D545BFABBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{11BB58D0-9592-424C-B9C9-71D545BFABBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{11BB58D0-9592-424C-B9C9-71D545BFABBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{11BB58D0-9592-424C-B9C9-71D545BFABBB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
