+++ examples/code-only/Example08_DebugShapes/Program.cs
@@ -1,4 +1,5 @@
 using Example08_DebugShapes.Scripts;
+using Stride.CommunityToolkit.DebugShapes.Code;
 using Stride.CommunityToolkit.Engine;
 using Stride.CommunityToolkit.Skyboxes;
 using Stride.Engine;
