+++ src/Stride.CommunityToolkit.DebugShapes/Code/ImmediateDebugPrimitives.cs
@@ -73,6 +73,7 @@
 // contributors exclude the implied warranties of merchantability, fitness for a
 // particular purpose and non-infringement.
 
+using Stride.Core.Mathematics;
 using Stride.Graphics;
 using Stride.Graphics.GeometricPrimitives;
