+++ src/Stride.CommunityToolkit.DebugShapes/Code/ImmediateDebugRenderFeature.cs
@@ -2,6 +2,7 @@
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using Stride.Core.Collections;
+using Stride.Core.Mathematics;
 using Stride.Core.Threading;
 using Stride.DebugRendering;
 using Stride.Graphics;
