+++ src/Stride.CommunityToolkit.DebugShapes/Code/ImmediateDebugRenderStageSelector.cs
