+++ src/Stride.CommunityToolkit.DebugShapes/Code/ImmediateDebugRenderSystem.cs
@@ -1,7 +1,9 @@
 // Copyright (c) Stride contributors (https://stride3d.net)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
 
+using Stride.Core;
 using Stride.Core.Collections;
+using Stride.Core.Mathematics;
 using Stride.Games;
 using Stride.Graphics;
 using Stride.Rendering;
