+++ src/Stride.CommunityToolkit.DebugShapes/Effects/LinePrimitiveShader.sdsl.cs
