+++ src/Stride.CommunityToolkit.DebugShapes/Effects/PrimitiveShader.sdsl.cs
