+++ src/Stride.CommunityToolkit.DebugShapes/Module.cs
@@ -1,0 +1,14 @@
+using Stride.Core;
+using Stride.Core.Reflection;
+using System.Reflection;
+
+namespace Stride.CommunityToolkit;
+
+internal class Module
+{
+    [ModuleInitializer]
+    public static void Initialize()
+    {
+        AssemblyRegistry.Register(typeof(Module).GetTypeInfo().Assembly, AssemblyCommonCategories.Assets);
+    }
+}
