+++ src/Stride.CommunityToolkit.DebugShapes/Stride.CommunityToolkit.DebugShapes.sdpkg
@@ -1,0 +1,4 @@
+!Package
+SerializedVersion: {Assets: 3.1.0.0}
+AssetFolders:
+    -   Path: !dir Effects
